backend/requirements.txt
fastapi
uvicorn
requests
pydantic
faiss-cpu
numpy
pypdf
python-multipart
google-genai
python-dotenv
Pillow
